bhyve: Move kernemu to amd64/
This code handles instruction emulation for accesses to various
amd64-specific MMIO regions.

No functional change intended.

Reviewed by:	corvink, jhb, emaste
MFC after:	1 week
Sponsored by:	Innovate UK
Differential Revision:	https://reviews.freebsd.org/D40554

(cherry picked from commit a1642451ce25bb0e2d01765a8dedda69c5029d48)
