+++ usr.sbin/bhyve/amd64/Makefile.inc
@@ -2,14 +2,16 @@ SRCS+=	\
 	atkbdc.c	\
 	e820.c		\
 	fwctl.c		\
+	kernemu_dev.c	\
 	mptbl.c		\
 	post.c		\
 	ps2kbd.c	\
 	ps2mouse.c	\
 	task_switch.c
 
-
 .PATH:  ${BHYVE_SYSDIR}/sys/amd64/vmm
 SRCS+=	vmm_instruction_emul.c
+
+CFLAGS.kernemu_dev.c+=	-I${SRCTOP}/sys/amd64
 
 SUBDIR+=	kbdlayout
